extract static methods from project regarding Actions to action.ts
